Move getStartOfMonth.js to luxon dir
